The following OS functions are now working: Kernel_OS_Init, Kernel_OS_Start, Kernel_Request_Handler, Kernel_Dispatch, Task_Next, Task_Create_System, Kernel_Task_Next, Kernel_Task_Create_System. Some of the listed are not in their final state, but are working for simple tests.
