+++ project_2/kernal.h
@@ -16,7 +16,6 @@
 /*
  * Type Definitions.
  */
-
 typedef void (*voidfuncptr) (void);	/* pointer to void f(void) */
 
 typedef enum process_states
@@ -59,28 +58,9 @@ typedef struct ProcessDescriptor
 	int creation_arg;
 	int pid;
 	unsigned int error;
+	PROCESS_PRIORITIES priority;
 } PD;
 
 /*
  * Structures.
  */
-
-/*
- * Functions.
- */
-void Kernel_OS_Abort(unsigned int error);
-// void exit_kernel(void);
-// void enter_kernel(void);
-
-PID  Kernel_Task_Create_System(void (*f)(void), int arg);
-PID  Kernel_Task_Create_RR(void (*f)(void), int arg);
-PID  Kernel_Task_Create_Period(void (*f)(void), int arg, TICK period, TICK wcet, TICK offset);
-void Kernel_Task_Next(void);
-int  Kernel_Task_GetArg(void);
-
-CHAN Kernel_Chan_Init();
-void Kernel_Send(CHAN ch, int v);
-int  Kernel_Recv(CHAN ch);
-void Kernel_Write(CHAN ch, int v);
-
-unsigned int Kernel_Now();
